refactor(hooks.useInfinityFetch): extract types to separated file
