+++ src/components/utils/Select/useSelectInfinityFetch/useInfinityFetch.ts
@@ -1,13 +1,6 @@
 import React from 'react';
 
-export type PaginationResponse<Response, Pagination> = {
-    response: Response;
-    pagination?: Pagination | null;
-};
-
-export type Fetcher<Response, Pagination> = (
-    pagination?: Pagination,
-) => Promise<PaginationResponse<Response, Pagination>>;
+import type {Fetcher, PaginationResponse} from './types';
 
 export const useInfinityFetch = <Response, Pagination>({
     fetcher,
